Perform a no-op if Node was not installed by the Heroku buildpack (#7)
On rare instances, a Heroku user may install their own version of Node
outside of the Node buildpack, ex: the Heroku CLI contains its own
version of Node which might not match the version installed via the
buildpack.

In this case, this version of Node would pick up the additional
--require statement in NODE_OPTIONS and include a native module
compiled for a different Node version and fail.

This commit adds some additional checks to make sure that the Node major
version matches the version the plugin was compiled for, and that the
Node executable is the one installed by the Node Buildpack. If either
is not true, this module becomes a no-op.